Reorganização da estrutura de arquivos e renomeação do diretório principal.
