--- ProjetoJSP/src/main/java/com/mycompany/testeprojetojsp/JakartaRestConfiguration.java
@@ -1,13 +1,0 @@
-package com.mycompany.testeprojetojsp;
-
-import jakarta.ws.rs.ApplicationPath;
-import jakarta.ws.rs.core.Application;
-
-/**
- * Configures Jakarta RESTful Web Services for the application.
- * @author Juneau
- */
-@ApplicationPath("resources")
-public class JakartaRestConfiguration extends Application {
-    
-}
