--- ProjetoJSP/src/main/java/com/mycompany/testeprojetojsp/resources/JakartaEE11Resource.java
@@ -1,20 +1,0 @@
-package com.mycompany.testeprojetojsp.resources;
-
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.core.Response;
-
-/**
- *
- * @author 
- */
-@Path("jakartaee11")
-public class JakartaEE11Resource {
-    
-    @GET
-    public Response ping(){
-        return Response
-                .ok("ping Jakarta EE")
-                .build();
-    }
-}
