+++ ProjetoJSP/src/main/java/model/Cliente.java
@@ -1,0 +1,5 @@
+package model;
+
+public class Cliente {
+    
+}
