+++ ProjetoJSP/src/main/java/resources/JakartaEE11Resource.java
@@ -1,0 +1,20 @@
+package resources;
+
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.core.Response;
+
+/**
+ *
+ * @author 
+ */
+@Path("jakartaee11")
+public class JakartaEE11Resource {
+    
+    @GET
+    public Response ping(){
+        return Response
+                .ok("ping Jakarta EE")
+                .build();
+    }
+}
